requirements.txt
streamlit
PyPDF2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
openpyxl